+++ src/mefit/Common/Constants/ApplicationVersions.cs
@@ -9,8 +9,7 @@ namespace Mac_EFI_Toolkit.Common.Constants
     internal static class ApplicationVersions
     {
         internal const string LZMA_SDK_VERSION = "24.08";
-        // YYmmDD.HHmm
-        internal const string CURRENT_BUILD = "250421.1550";
-        internal const string CURRENT_CHANNEL = "Stable";
+        internal const string BUILD = "20250421_1550"; // yyyyMMdd_HHmm
+        internal const string CHANNEL = "Stable";
     }
 }
