+++ src/mefit/Forms/frmAbout.cs
@@ -42,8 +42,8 @@ namespace Mac_EFI_Toolkit.Forms
         {
             lblBuild.Text =
                 $"Mac EFI Toolkit v{Application.ProductVersion} by Muerto{Environment.NewLine}" +
-                $"Build {ApplicationVersions.CURRENT_BUILD} ({ApplicationVersions.CURRENT_CHANNEL}){Environment.NewLine}" +
-                $"LZMA SDK {ApplicationVersions.LZMA_SDK_VERSION}";
+                $"Build: {ApplicationVersions.BUILD} ({ApplicationVersions.CHANNEL}){Environment.NewLine}" +
+                $"LZMA SDK: v{ApplicationVersions.LZMA_SDK_VERSION}";
         }
         #endregion
