+++ src/mefit/Tools/SystemTools.cs
@@ -4,6 +4,7 @@
 // SystemTools.cs
 // Released under the GNU GLP v3.0
 
+using Mac_EFI_Toolkit.Common.Constants;
 using Mac_EFI_Toolkit.WIN32;
 using System;
 using System.Diagnostics;
