+++ src/mefit/UI/UITools.cs
@@ -5,6 +5,7 @@
 // UITools.cs
 // Released under the GNU GLP v3.0
 
+using Mac_EFI_Toolkit.Common.Constants;
 using Mac_EFI_Toolkit.WIN32;
 using System;
 using System.Diagnostics;
